Helper function moved
