+++ aasrp/utils.py
@@ -9,7 +9,6 @@ from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 
 # Alliance Auth
-from allianceauth.authentication.admin import User
 from allianceauth.services.hooks import get_extension_logger
 
 # Alliance Auth (External Libs)
@@ -72,23 +71,3 @@ def clean_setting(
 DATETIME_FORMAT = "%Y-%m-%d %H:%M"
 
 format_html_lazy = lazy(format_html, str)
-
-
-def get_main_character_from_user(user: User) -> str:
-    """
-    Get the main character from a user
-    :param user:
-    :type user:
-    :return:
-    :rtype:
-    """
-
-    user_main_character = user.username
-
-    try:
-        user_profile = user.profile
-        user_main_character = user_profile.main_character.character_name
-    except AttributeError:
-        pass
-
-    return user_main_character
